feat: add xtts narration provider

+++ adapters/sqlite/migrations/00001_init.sql
@@ -176,8 +176,8 @@ CREATE TABLE settings (
     type        TEXT    NOT NULL,
     grp         TEXT    NOT NULL DEFAULT '',
     description TEXT    NOT NULL DEFAULT '',
-    min_value   INTEGER NOT NULL DEFAULT 0,
-    max_value   INTEGER NOT NULL DEFAULT 0,
+    min_value   REAL    NOT NULL DEFAULT 0,
+    max_value   REAL    NOT NULL DEFAULT 0,
     updated_at  INTEGER NOT NULL
 );
