+++ adapters/sqlite/sqlcgen/models.go
@@ -50,8 +50,8 @@ type Setting struct {
 	Type        string
 	Grp         string
 	Description string
-	MinValue    int64
-	MaxValue    int64
+	MinValue    float64
+	MaxValue    float64
 	UpdatedAt   int64
 }
